requirements.txt
jupyterlab
matplotlib
plotly
ipywidgets
jupyterlab_widgets
pandas
pymongo
python-binance
tqdm